+++ README.adoc
@@ -8,11 +8,11 @@ https://www.codacy.com/app/damith/addressbook-level3?utm_source=github.com&utm_m
 
 
 ifdef::env-github[]
-image::docs/images/Ui.png[width="600"]
+image::docs/images/View.png[width="600"]
 endif::[]
 
 ifndef::env-github[]
-image::images/Ui.png[width="600"]
+image::images/View.png[width="600"]
 endif::[]
 
 * This is a desktop Address Book application. It has a GUI but most of the user interactions happen using a CLI (Command Line Interface).
