+++ docs/UsingPlantUml.adoc
@@ -95,7 +95,7 @@ For a comprehensive list of ``skinparam``s head over to the https://plantuml-doc
 While PlantUML's automatic layout engine usually produces satisfactory results, at times the result can be less than ideal, especially on larger diagrams.
 Here is an example where the default layout generated by PlantUML has a lot of overlapping lines that are hard to decipher:
 
-.The UI class diagram without additional formatting
+.The VIEW class diagram without additional formatting
 image::RawUiDiagram.png[]
 
 NOTE: In most cases, you should consider decomposing the diagram into smaller ones or focusing on a more specific portion of the diagram.
