+++ src/main/java/seedu/address/MainApp.java
@@ -28,8 +28,8 @@ import seedu.address.storage.JsonUserPrefsStorage;
 import seedu.address.storage.Storage;
 import seedu.address.storage.StorageManager;
 import seedu.address.storage.UserPrefsStorage;
-import seedu.address.ui.Ui;
-import seedu.address.ui.UiManager;
+import seedu.address.view.View;
+import seedu.address.view.UiManager;
 
 /**
  * Runs the application.
@@ -40,7 +40,7 @@ public class MainApp extends Application {
 
     private static final Logger logger = LogsCenter.getLogger(MainApp.class);
 
-    protected Ui ui;
+    protected View view;
     protected Logic logic;
     protected Storage storage;
     protected Model model;
@@ -65,7 +65,7 @@ public class MainApp extends Application {
 
         logic = new LogicManager(model, storage);
 
-        ui = new UiManager(logic);
+        view = new UiManager(logic);
     }
 
     /**
@@ -168,7 +168,7 @@ public class MainApp extends Application {
     @Override
     public void start(Stage primaryStage) {
         logger.info("Starting AddressBook " + MainApp.VERSION);
-        ui.start(primaryStage);
+        view.start(primaryStage);
     }
 
     @Override
