--- src/main/java/seedu/address/ui/Ui.java
@@ -1,13 +1,0 @@
-package seedu.address.ui;
-
-import javafx.stage.Stage;
-
-/**
- * API of UI component
- */
-public interface Ui {
-
-    /** Starts the UI (and the App).  */
-    void start(Stage primaryStage);
-
-}
