+++ src/main/java/seedu/address/view/ResultDisplay.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.view;
 
 import static java.util.Objects.requireNonNull;
 
@@ -7,7 +7,7 @@ import javafx.scene.control.TextArea;
 import javafx.scene.layout.Region;
 
 /**
- * A ui for the status bar that is displayed at the header of the application.
+ * A view for the status bar that is displayed at the header of the application.
  */
 public class ResultDisplay extends UiPart<Region> {
