+++ src/main/java/seedu/address/view/UiPart.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.view;
 
 import static java.util.Objects.requireNonNull;
 
@@ -9,7 +9,7 @@ import javafx.fxml.FXMLLoader;
 import seedu.address.MainApp;
 
 /**
- * Represents a distinct part of the UI. e.g. Windows, dialogs, panels, status bars, etc.
+ * Represents a distinct part of the VIEW. e.g. Windows, dialogs, panels, status bars, etc.
  * It contains a scene graph with a root node of type {@code T}.
  */
 public abstract class UiPart<T> {
