+++ src/main/java/seedu/address/view/View.java
@@ -1,0 +1,13 @@
+package seedu.address.view;
+
+import javafx.stage.Stage;
+
+/**
+ * API of VIEW component
+ */
+public interface View {
+
+    /** Starts the VIEW (and the App).  */
+    void start(Stage primaryStage);
+
+}
