+++ src/test/resources/view/UiPartTest/validFile.fxml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import seedu.address.ui.TestFxmlObject?>
+<?import seedu.address.view.TestFxmlObject?>
 <TestFxmlObject xmlns:fx="http://javafx.com/fxml/1" fx:id="validFileRoot">Hello World!</TestFxmlObject>
